+++ ZeroClipboard.js
@@ -121,7 +121,6 @@
   ZeroClipboard.Client.prototype.setText = function(newText) {
     if (newText && newText !== "") {
       this._text = newText;
-      this.htmlBridge.setAttribute("data-clipboard-text", newText);
       if (this.ready()) this.flashBridge.setText(newText);
     }
   };
@@ -204,6 +203,9 @@
       break;
      case "mouseup":
       ZeroClipboard.currentElement.removeClass("zeroclipboard-is-active");
+      break;
+     case "complete":
+      this.resetText();
       break;
     }
     if (this.handlers[eventName]) {