+++ docs/releases.md
@@ -6,6 +6,8 @@ This release re-writes a lot of how zeroclipboard client works. Check the docs f
 * [FEATURE] Indicate via an event when there is no flash or flash is wrong version
 * [FEATURE] More meta info passed with each flash event.
 * [FEATURE] Register more than one event `.on("load complete", function() {})`
+* [FEATURE] Event handlers can now access the clicked element via `this`
+* [BUG] When creating more than one clip, the second clip doesn't glue.
 
 ### ZeroClipboard 1.0.9